requirements.txt
fastapi
uvicorn
pymongo
python-multipart
jinja2
requests
websockets
motor
python-dotenv
pytz